feat: refactor season structure and update namespaces
- move season-related classes to a modular structure
- update ns for SeasonFactory, SeasonRepository, and related classes
- add season routes for retrieving current and all seasons
