+++ app-modules/season/src/Models/Season.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Heart\Season\Infrastructure\Models;
+namespace He4rt\Season\Models;
 
-use Heart\Badges\Infrastructure\Model\Badge;
+use He4rt\Badge\Models\Badge;
+use He4rt\Season\Database\Factories\SeasonFactory;
 use Heart\Meeting\Infrastructure\Models\Meeting;
-use Heart\Season\Infrastructure\Factories\SeasonFactory;
 use Illuminate\Database\Eloquent\Concerns\HasUuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
